# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ManagemenLogbook API",
        "description": "API documentation for ManagemenLogbook",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login User menggunakan Email atau Username",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "login_identity",
                                    "password"
                                ],
                                "properties": {
                                    "login_identity": {
                                        "type": "string",
                                        "example": "rasyid_ridho atau rasyid@polman.astra.ac.id"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login Berhasil"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/api/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout User",
                "operationId": "ad65cbbd4e9f201619eae184a5961a98",
                "responses": {
                    "200": {
                        "description": "Logout Berhasil"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/logbook": {
            "get": {
                "tags": [
                    "Logbook"
                ],
                "summary": "Get logbooks (Filter Task/Date/Search)",
                "operationId": "6b99b1d92b104d2bb3fc9536aa410471",
                "parameters": [
                    {
                        "name": "tgs_id",
                        "in": "query",
                        "description": "Filter Task ID",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "tanggal",
                        "in": "query",
                        "description": "Filter Date",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search description",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of logbook entries"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Logbook"
                ],
                "summary": "Create logbook entry",
                "operationId": "b9315d9ece6c77a5f31a4a31021da052",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "tgs_id",
                                    "tanggal",
                                    "deskripsi"
                                ],
                                "properties": {
                                    "tgs_id": {
                                        "type": "integer"
                                    },
                                    "tanggal": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "deskripsi": {
                                        "type": "string"
                                    },
                                    "komentar": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Logbook entry created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}